requirements.txt
Django~=2.1.7
pip~=20.2.1
wheel~=0.36.2
pytz~=2020.5
sqlparse~=0.4.1
asgiref~=3.3.1
signals~=0.0.2
setuptools~=49.2.1
django-crispy-forms~=1.11.0
Pillow~=8.1.0
django-ckeditor==6.0.0
urllib3==1.26.2
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
django-js-asset==1.2.2
sendgrid